Esp32c3 wakeup ext (#1)
* modmachine: Implemented GPIO wake-up for ESP32-C3.

As ESP32-C3 uses GPIO wake-up instead of EXT0/1 wake-up,
several small changes in code was required. Interface stays
the same as for generic ESP32 but EXT0 is ignored and EXT1
is used as wake-up source.

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>

* modmachine: Fixed wake-up reason for on ESP32-C3.

Wake-up reason for pin wake-up is not EXT1 but GPIO on ESP32-C3.

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>

* modmachine, modesp32: Fixed wake-up RTC pins chck.

There have been incorrect test of used EXT1 (GPIO) wake-up capable
(RTC) pins for ESP32-C3.

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>

* modmachine, modesp32: Fixed deep/light sleep.

With ESP32-C3 there is different approach for setting wake-up pins.

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>

* modmachine: ESP32 deepsleep on RTC pins only.

The light sleep on ESP32-C3 is allowed on all IO pins, however deepsleep
on RTC pins only. This patch raises exception when deepsleep is invoked
and non-RTC pins are used as wake-up sources.

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>

---------

Signed-off-by: Ondrej Sienczak <ondrej.sienczak@gmail.com>
Co-authored-by: OSi (Ondrej Sienczak) <ondrej.sienczak@seznam.cz>